add output format control
