--- benchmark/benchmark_responses.json
@@ -1,17 +1,0 @@
-[
-    {
-        "image": "https://s2.loli.net/2023/05/17/OyhFvmaH867DwC2.jpg",
-        "prompt": "Is it in real scene?",
-        "response": "<s> <image> User: Is it in real scene? GPT: <answer> The image depicts a man standing in a city setting with tall buildings and a yellow fire hydrant in the foreground. It is not a computer-generated image or a graphic novel scene, but rather a real-world urban environment. <|endofchunk|>"
-    },
-    {
-        "image": "https://s2.loli.net/2023/05/17/7OaJlqyRri2E8zY.png",
-        "prompt": "What's written on this image?",
-        "response": "<s> <image> User: What's written on this image? GPT: <answer> \"The image features a close-up of a Dove bottle with the word \"\"Dove\"\" written on it.\". <|endofchunk|>"
-    },
-    {
-        "image": "https://s2.loli.net/2023/05/17/18egkN6aYGDAtsM.png",
-        "prompt": "Which city is this?",
-        "response": "<s> <image> User: Which city is this? GPT: <answer> The image shows the city of Singapore. <|endofchunk|>"
-    }
-]
