+++ evaluation/sample_questions.json
@@ -1,0 +1,14 @@
+[
+    {
+        "image": "https://s2.loli.net/2023/05/17/OyhFvmaH867DwC2.jpg",
+        "prompt": "Is this scene real?"
+    },
+    {
+        "image": "https://s2.loli.net/2023/05/17/7OaJlqyRri2E8zY.png",
+        "prompt": "What's written on this image?"
+    },
+    {
+        "image": "https://s2.loli.net/2023/05/17/18egkN6aYGDAtsM.png",
+        "prompt": "Which city is this?"
+    }
+]
